Organized code for calendar page
- moved upcoming events section on to a separate component file
- updated banner in the calendar page
